Give flags as unsigned long to mysql_real_connect()
Converting between int and unsigned long can cause extra bits to
be set. Older versions of Connector-C don't check the flags, but
more recent versions do.

+++ system-test/maxtest/include/maxtest/mariadb_func.hh
@@ -485,5 +485,6 @@ private:
     bool        m_ssl;
     int         m_timeout = 0;
     MYSQL*      m_conn = nullptr;
-    int         m_options = CLIENT_MULTI_STATEMENTS;
+
+    unsigned long m_options = CLIENT_MULTI_STATEMENTS;
 };
